% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{blueterra_options}
\alias{blueterra_options}
\title{Configure blueterra runtime options}
\usage{
blueterra_options(...)
}
\arguments{
\item{...}{Named option values. Currently supported options are
\code{blueterra.progress} and \code{blueterra.max_plot_cells}.}
}
\value{
A named list with current option values, invisibly when setting.
}
\description{
Returns or sets package options used by examples and helper functions.
}
\details{
Options affect only local package behavior and do not write outside paths
provided by the user.
}
\examples{
blueterra_options()
old <- blueterra_options(blueterra.progress = FALSE)
blueterra_options(blueterra.progress = old$blueterra.progress)

}
