% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{crop_bathy}
\alias{crop_bathy}
\alias{mask_bathy}
\alias{resample_bathy}
\alias{project_bathy}
\title{Crop, mask, resample, and project bathymetry}
\usage{
crop_bathy(x, extent, filename = "", overwrite = FALSE)

mask_bathy(x, mask, filename = "", overwrite = FALSE)

resample_bathy(x, y, method = "bilinear", filename = "", overwrite = FALSE)

project_bathy(x, crs, method = "bilinear", filename = "", overwrite = FALSE)
}
\arguments{
\item{x}{A raster-like object accepted by \code{\link[=as_bathy]{as_bathy()}}.}

\item{extent}{Crop extent.}

\item{filename}{Optional output raster path.}

\item{overwrite}{Logical. Allow overwriting \code{filename}.}

\item{mask}{Polygon/vector mask.}

\item{y}{Template raster used for resampling.}

\item{method}{Interpolation method passed to \code{terra}.}

\item{crs}{Target CRS.}
}
\value{
A \code{terra::SpatRaster}.
}
\description{
Small wrappers around \code{terra} raster-preparation operations with bathymetry
input validation.
}
\details{
These helpers preserve depth sign and raster values except for interpolation
effects introduced by resampling or projection.
}
\examples{
bathy <- read_bathy(blueterra_example("bathy"))
zones <- terra::vect(blueterra_example("zones"))
crop_bathy(bathy, terra::ext(zones[zones$site_id == "slope", ]))

}
\seealso{
\code{\link[=prepare_bathy]{prepare_bathy()}}
}
