% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isobath-corridors.R
\name{plot_isobath_corridors}
\alias{plot_isobath_corridors}
\title{Plot isobath corridors}
\usage{
plot_isobath_corridors(corridors, bathy = NULL)
}
\arguments{
\item{corridors}{Corridor polygons.}

\item{bathy}{Optional raster background.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots corridor polygons, optionally over a raster background.
}
\examples{
bathy <- read_bathy(blueterra_example("bathy"))
corridors <- make_isobath_corridors(bathy, depths = -50, width = 20)
plot_isobath_corridors(corridors, bathy)

}
\seealso{
\code{\link[=make_isobath_corridors]{make_isobath_corridors()}}
}
