% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_process_pca}
\alias{plot_process_pca}
\title{Plot terrain PCA}
\usage{
plot_process_pca(pca)
}
\arguments{
\item{pca}{Output from \code{\link[=terrain_pca]{terrain_pca()}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots the first two principal component score axes from \code{\link[=terrain_pca]{terrain_pca()}}.
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  df <- data.frame(a = rnorm(20), b = rnorm(20), c = rnorm(20))
  plot_process_pca(terrain_pca(df))
}

}
\seealso{
\code{\link[=terrain_pca]{terrain_pca()}}
}
